Seguimos trabajando , todavia no he terminado los modders

+++ src/lab2progra2_carlosnoé/Casas.java
@@ -57,7 +57,7 @@ private String Estado;
         return largo;
     }
 
-    public void setLargo(int largo) {
+    public void setLargo(double largo) {
         this.largo = largo;
     }
